% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BioneerV2}
\alias{BioneerV2}
\title{Bioneer Library Version 2}
\format{An object of class \code{data.frame} with 3006 rows and 4 columns.}
\usage{
BioneerV2
}
\seealso{
\link{BioneerV5}
}
\keyword{datasets}
