% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BioneerV5}
\alias{BioneerV5}
\title{Bioneer Library Version 5}
\format{An object of class \code{data.frame} with 3420 rows and 6 columns.}
\usage{
BioneerV5
}
\seealso{
\link{BioneerV2}
}
\keyword{datasets}
