% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Safe_Barcodes.R
\name{calculateFrequency}
\alias{calculateFrequency}
\alias{aggregateProportions}
\title{Calculation of Barcode Frequencies}
\usage{
calculateFrequency(barcodeTable, barcodeColumn, geneColumn)

aggregateProportions(barcodeFrequencyTable, barcodeColumn, geneColumn)
}
\arguments{
\item{barcodeTable}{A \code{data.frame} containing the barcode and gene data}

\item{barcodeColumn}{The column containing the barcodes}

\item{geneColumn}{The column containing the genes}

\item{barcodeFrequencyTable}{A \code{data.frame} as produced by \code{calculateFrequency}.}
}
\description{
\code{calculateFrequency} will calculate the total number of times each barcode-gene combination pair appears.

\code{aggregateProportions} will calculate the best match of genes to barcodes and provide information of the number of matches, mismatches, total matches and proportion of matches.
}
