% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_Name_Manipulation.R
\name{common2systematic}
\alias{common2systematic}
\title{Conversion of common/alias names to systematic IDs}
\usage{
common2systematic(common)
}
\arguments{
\item{common}{A character vector of the names to be checked}
}
\value{
A character vector the same length as the input is returned with updated IDs.
In the case that a \code{NULL} or empty string is provided, \code{"NA"} will  be returned.
In the case that no match is found in current names, the name will be checked to see if it has been removed, in which case \code{"Gene Removed"} will be returned.
If this also fails, \code{"Gene Not Found"} will be returned.
}
