% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barcode_Matching.R
\name{createConsensus}
\alias{createConsensus}
\title{Create consensus barcodes from a vector of barcodes}
\usage{
createConsensus(barcodes, maxMismatch = 1, clusterPlot = NULL,
  plotWidth = 1000)
}
\arguments{
\item{barcodes}{A character vector containing the barcodes}

\item{clusterPlot}{Should a dendrogram of the clustering be plotted?
In the case of \code{NULL} (the default), there will be no plot.
In order to save a dendrogram, please specify a file path here.
This should be a pdf.}

\item{plotWidth}{The plot width of the dendrogram (this is usuall very large).}

\item{maxMistmatch}{The maximum number of mismatches which will be allowed to call two barcodes the same}
}
\value{
A \code{data.frame} is returned showing the consensus barcodes and how many times a barcode has been matched to the consensus
Defauts to 3
}
