% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDTOOLS_Import.R
\name{ExtractBarcodefromBEDTOOL}
\alias{ExtractBarcodefromBEDTOOL}
\title{Importing data from bedtools}
\usage{
ExtractBarcodefromBEDTOOL(file, outputfile)
}
\arguments{
\item{file}{The file path to the bedtools output.}

\item{outputFile}{The file path to the csv to be created.}
}
\value{
Nothing is returned from this function. Instead, a new csv file is written.
}
\description{
\code{ExtractBarcodefromBEDTOOL} will read a file outputted from bedtools, extract the barcode information and remove unnecessary columns.
}
