% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barcode_Matching.R
\name{matchBarcodes}
\alias{matchBarcodes}
\title{Matching Barcodes to a Barcode Library}
\usage{
matchBarcodes(barcodeTable, barcodeColumn, barcodeLibrary)
}
\arguments{
\item{barcodeTable}{A \code{data.frame} containing the barcodes to be matched}

\item{barcodeColumn}{Which column in \code{barcodeTable} contains the barcodes?}

\item{barcodeLibrary}{A vector containing the barcodes to be matched against}
}
\value{
\code{matchBarcodes} returns the original \code{data.frame} supplied to the function, but with two additional columns containing the best matched barcode and the number of mismatches between the supplied barcode and the best match.
}
