% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pombeGenes}
\alias{pombeGenes}
\title{Current S. pombe Genes}
\format{An object of class \code{data.frame} with 5138 rows and 4 columns.}
\source{
\url{ftp://ftp.pombase.org/pombe/names_and_identifiers/sysID2product.tsv}
}
\usage{
pombeGenes
}
\description{
A table showing current S. pombe systematic IDs, common names and aliases
}
\seealso{
\link{pombeRemovedGenes}
}
\keyword{datasets}
