% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pombeRemovedGenes}
\alias{pombeRemovedGenes}
\title{Removed S. pombe Genes}
\format{An object of class \code{data.frame} with 80 rows and 3 columns.}
\source{
\url{https://www.pombase.org/status/new-and-removed-genes}
}
\usage{
pombeRemovedGenes
}
\description{
A table showing removed S. pombe systematic IDs
}
\seealso{
\link{pombeGenes}
}
\keyword{datasets}
